refactor: extract get_param_as_byte function for improved parameter handling and fix backup inconsistencies
